sql/mon, util/mon: move bytes_usage to a util package
