+++ pkg/sql/distsqlrun/distinct.go
@@ -17,9 +17,9 @@ package distsqlrun
 import (
 	"sync"
 
-	"github.com/cockroachdb/cockroach/pkg/sql/mon"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlbase"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
+	"github.com/cockroachdb/cockroach/pkg/util/mon"
 	"github.com/cockroachdb/cockroach/pkg/util/tracing"
 	"golang.org/x/net/context"
 )
