+++ pkg/sql/parser/aggregate_builtins.go
@@ -22,9 +22,9 @@ import (
 	"golang.org/x/net/context"
 
 	"github.com/cockroachdb/apd"
-	"github.com/cockroachdb/cockroach/pkg/sql/mon"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
 	"github.com/cockroachdb/cockroach/pkg/util/duration"
+	"github.com/cockroachdb/cockroach/pkg/util/mon"
 )
 
 func initAggregateBuiltins() {
